+++ jsonpath4k/src/commonMain/kotlin/at/asitplus/jsonpath/core/JsonPathFunctionExtension.kt
@@ -7,18 +7,21 @@ import kotlinx.serialization.json.JsonElement
  * date: 2024-02
  * section: 2.4.  Function Extensions
  */
-sealed class JsonPathFunctionExtension<ReturnType : JsonPathFilterExpressionValue>(
+sealed class JsonPathFunctionExtension(
     vararg val argumentTypes: JsonPathFilterExpressionType,
 ) {
-    abstract fun evaluate(arguments: List<JsonPathFilterExpressionValue>): ReturnType
+    abstract operator fun invoke(arguments: List<JsonPathFilterExpressionValue>): JsonPathFilterExpressionValue
+
+    @Deprecated("To be removed in version 4. Replaced with `invoke`.", ReplaceWith("invoke"))
+    fun evaluate(arguments: List<JsonPathFilterExpressionValue>): JsonPathFilterExpressionValue = invoke(arguments)
 
     class ValueTypeFunctionExtension(
         vararg argumentTypes: JsonPathFilterExpressionType,
-        private val evaluator: (arguments: List<JsonPathFilterExpressionValue>) -> JsonElement?
-    ) : JsonPathFunctionExtension<JsonPathFilterExpressionValue.ValueTypeValue>(
+        private val evaluator: (arguments: List<JsonPathFilterExpressionValue>) -> JsonElement?,
+    ) : JsonPathFunctionExtension(
         argumentTypes = argumentTypes,
     ) {
-        override fun evaluate(arguments: List<JsonPathFilterExpressionValue>): JsonPathFilterExpressionValue.ValueTypeValue {
+        override operator fun invoke(arguments: List<JsonPathFilterExpressionValue>): JsonPathFilterExpressionValue.ValueTypeValue {
             return evaluator(arguments)?.let {
                 JsonPathFilterExpressionValue.ValueTypeValue.JsonValue(it)
             } ?: JsonPathFilterExpressionValue.ValueTypeValue.Nothing
@@ -27,22 +30,22 @@ sealed class JsonPathFunctionExtension<ReturnType : JsonPathFilterExpressionValu
 
     class LogicalTypeFunctionExtension(
         vararg argumentTypes: JsonPathFilterExpressionType,
-        private val evaluator: (arguments: List<JsonPathFilterExpressionValue>) -> Boolean
-    ) : JsonPathFunctionExtension<JsonPathFilterExpressionValue.LogicalTypeValue>(
+        private val evaluator: (arguments: List<JsonPathFilterExpressionValue>) -> Boolean,
+    ) : JsonPathFunctionExtension(
         argumentTypes = argumentTypes,
     ) {
-        override fun evaluate(arguments: List<JsonPathFilterExpressionValue>): JsonPathFilterExpressionValue.LogicalTypeValue {
+        override operator fun invoke(arguments: List<JsonPathFilterExpressionValue>): JsonPathFilterExpressionValue.LogicalTypeValue {
             return JsonPathFilterExpressionValue.LogicalTypeValue(evaluator(arguments))
         }
     }
 
     class NodesTypeFunctionExtension(
         vararg argumentTypes: JsonPathFilterExpressionType,
-        private val evaluator: (arguments: List<JsonPathFilterExpressionValue>) -> List<JsonElement>
-    ) : JsonPathFunctionExtension<JsonPathFilterExpressionValue.NodesTypeValue.FunctionExtensionResult>(
+        private val evaluator: (arguments: List<JsonPathFilterExpressionValue>) -> List<JsonElement>,
+    ) : JsonPathFunctionExtension(
         argumentTypes = argumentTypes,
     ) {
-        override fun evaluate(arguments: List<JsonPathFilterExpressionValue>): JsonPathFilterExpressionValue.NodesTypeValue.FunctionExtensionResult {
+        override operator fun invoke(arguments: List<JsonPathFilterExpressionValue>): JsonPathFilterExpressionValue.NodesTypeValue.FunctionExtensionResult {
             return JsonPathFilterExpressionValue.NodesTypeValue.FunctionExtensionResult(
                 evaluator(arguments)
             )